fix: rename file to index.jsx, fixed styling
