+++ scrybe_frontend/src/pages/HelpAndSupport/components/General1/General1.module.scss
@@ -4,7 +4,14 @@
         font-weight: 600;
         font-size: 20px !important;
     }
+       @media screen and (max-width: 1240px) {
+    
+        padding: 0 1rem;
+    
 }
+}
+
+
 
 
 .articles_cont {
@@ -36,6 +43,14 @@
         border-bottom: transparent;
     }
 }
+
+@media screen and (max-width: 1240px) {
+.general1 {
+padding: 0 1rem;
+}
+}
+
+
 
 @media screen and (max-width: 900px) {
     .info_intro {
