+++ scrybe_frontend/src/pages/HelpAndSupport/components/General1/index.jsx
@@ -3,9 +3,9 @@ import Articles from "../articles/Articles";
 import PropTypes from "prop-types";
 function General1({ setDisplay }) {
   return (
-    <div>
+    <div className={styles.general1}>
       <div className={styles.info_intro}>
-        <p>General Info On Scrybe</p>
+        <p>General Info On Heed</p>
       </div>
 
       <div className={styles.articles_cont}>
@@ -13,7 +13,7 @@ function General1({ setDisplay }) {
           <h3 onClick={() => setDisplay(false)}>
             Guide: How to log in and sign up
           </h3>
-          <h3>Guide: Who can use scrybe?</h3>
+          <h3>Guide: Who can use Heed?</h3>
           <h3>Guide: How Accurate is the Transcriptor? </h3>
           <h3>Guide: Monitoring your employees data </h3>
           <h3>Article: How to get your downloads in real time</h3>
